% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{flanker_data}
\alias{flanker_data}
\title{Example response time data set from a flanker paradigm.}
\format{
A data frame with 12524 rows and 5 variables:
\describe{
\item{subject}{The subject identification number. Note that some
participants were removed due to poor accuracy.}
\item{warning}{The experimental condition (2 in this example). It
logs the presence/absence of an auditory warning tone before
stimulus onset.}
\item{congruency}{The congruency of the flanker stimulus.}
\item{accuracy}{Accuracy of the response; 1 = correct, 0 = error}
\item{rt}{Response time, recorded in seconds.}
}
#' @references
Grange, J.A. (2016). flankr: An R package for implementing computational
models of attentional selectivity. Behavior Research Methods, 48, 528–541.
https://doi.org/10.3758/s13428-015-0615-y
}
\usage{
flanker_data
}
\description{
An example data set from an Eriksen flanker task containing multiple
participants' data. As well as congruency of the flanker stimuli, the
experiment had an additional condition of "warning", which was whether a
warning tone was played before stimulus onset or not. Response time and
accuracy is recorded for each trial.
}
\keyword{datasets}
