% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdf.R
\name{get_cdf}
\alias{get_cdf}
\title{Compute Conditional Distribution Functions (CDF) from Response Time Data}
\usage{
get_cdf(
  data,
  rt_var = "rt",
  accuracy_var = "accuracy",
  id_var = NULL,
  conditions = NULL,
  include_errors = FALSE,
  probs = c(0.1, 0.3, 0.5, 0.7, 0.9)
)
}
\arguments{
\item{data}{A data frame / tibble in long format with columns containing
(at the very least) response time and accuracy.}

\item{rt_var}{The column name where response times can be found.}

\item{accuracy_var}{The column name where accuracy can be found.}

\item{id_var}{If the data contains responses from multiple participants, the
column name where the participant id can be found.}

\item{conditions}{If the data contains responses from multiple conditions,
list the column names where the levels of the condition are coded.
If there are multiple conditions, provide a concatenated vector of condition
column names (e.g., c("age_group", "gender")).

Note that for plotting, only three conditions can be plotted. The order of conditions matters when there
are more than two conditions:
\itemize{
\item If there are just two conditions,the first condition will be grouped on the plot and the second condition will
be faceted as columns.
\item If there are three conditions, the first condition
will be grouped, the second will be faceted as columns, and the third will
be faceted as rows.
}}

\item{include_errors}{A logical variable (TRUE/FALSE) indicating whether the
CDF calculations should include error trials or not.}

\item{probs}{A vector of probability values for the CDF.}
}
\description{
This function computes CDF values from response time data, optionally grouping by conditions.
}
