% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Opposite of \%in\%: Not-In Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A vector of values to be tested.}

\item{table}{A vector, matrix, or data frame to be matched against.
Each element of \code{x} is tested for absence from this object.}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Returns a logical vector indicating if elements of the left-hand vector (\code{x})
are not present in the right-hand vector (\code{table}). This is the logical negation of \code{\%in\%}.
}
