% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{task_switching}
\alias{task_switching}
\title{Sample data from a task switching paradigm by Grange (2025)}
\format{
A data frame with 24,376 rows and 6 variables:
\describe{
\item{id}{participant identification}
\item{response_set_sequence}{whether the current trial is a repeat or a
switch of response set}
\item{stimulus_set_sequence}{whether the current trial is a repeat or a
switch of stimulus set}
\item{compatibility}{whether the current trial's response is compatible
or incompatible}
\item{rt}{response time in milliseconds}
\item{accuracy}{accuracy of the response: 1 for correct, 0 for error}
}
}
\usage{
task_switching
}
\description{
A data set including data from 51 participants from a task switching study.
Experimental manipulations included stimulus set sequence (repetition vs.
switch of which stimulus to attend to), response set sequence (repetition
vs. switch of what task to perform on the attended stimulus), and
compatibility (whether the responses are compatible or incompatible).
Response time and accuracy is recorded for each trial.
}
\references{
Grange, J.A. (2025). Control of stimulus set and response set in task
switching. Journal of Experimental Psychology: Learning, Memory, &
Cognition. https://doi.org/10.1037/xlm0001459
}
\keyword{datasets}
